backend/admin/requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
